+++ src/main/java/com/website/models/CartEntity.java
@@ -10,10 +10,10 @@ public class CartEntity {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "total_price")
-    private Long totalPrice;
+    private Integer totalPrice;
 
     @Column(name = "status")
     private Integer status;
@@ -38,19 +38,19 @@ public class CartEntity {
     public CartEntity() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getTotalPrice() {
+    public Integer getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Long totalPrice) {
+    public void setTotalPrice(Integer totalPrice) {
         this.totalPrice = totalPrice;
     }
