+++ src/main/java/com/website/models/Product_GoodReceived_Key.java
@@ -9,32 +9,32 @@ import java.util.Objects;
 public class Product_GoodReceived_Key implements Serializable {
 
     @Column(name = "goodsreceived_id")
-    private Long goodsreceived_id;
+    private Integer goodsreceived_id;
 
     @Column(name = "pro_id")
-    private Long pro_id;
+    private Integer pro_id;
 
     public Product_GoodReceived_Key() {
     }
 
-    public Product_GoodReceived_Key(Long goodsreceived_id, Long pro_id) {
+    public Product_GoodReceived_Key(Integer goodsreceived_id, Integer pro_id) {
         this.goodsreceived_id = goodsreceived_id;
         this.pro_id = pro_id;
     }
 
-    public Long getGoodsreceived_id() {
+    public Integer getGoodsreceived_id() {
         return goodsreceived_id;
     }
 
-    public void setGoodsreceived_id(Long goodsreceived_id) {
+    public void setGoodsreceived_id(Integer goodsreceived_id) {
         this.goodsreceived_id = goodsreceived_id;
     }
 
-    public Long getPro_id() {
+    public Integer getPro_id() {
         return pro_id;
     }
 
-    public void setPro_id(Long pro_id) {
+    public void setPro_id(Integer pro_id) {
         this.pro_id = pro_id;
     }
